backend/requirements.txt
flask
flask-cors
python-dotenv
requests
PyPDF2
google-cloud-documentai
google-cloud-aiplatform
